pyproject.toml
[tool.poetry]
name = "linkedin-job-scraper"
version = "0.1.0"
description = "Having some fun with LinkedIn data"
authors = ["Fernando Cordeiro <email@gmail.com>"]
readme = "README.md"
packages = [{include = "job_scraper"}]

[tool.poetry.dependencies]
python = "^3.11"
requests = "^2.28.2"
python-dotenv = "^1.0.0"
selenium = "^4.8.2"
python-slugify = "^8.0.1"
beautifulsoup4 = "^4.12.0"
fake-useragent = "^1.1.1"
sqlalchemy = "^2.0.7"
click = "^8.1.3"

[tool.poetry.group.dev.dependencies]
# Linters
black = {extras = ["jupyter"], version = "^23.1.0"}
isort = "^5.12.0"
pylint = "^2.17.0"
# Testing
pytest = "^7.2.2"
# Notebooks
ipykernel = "^6.21.3"
pandas = "^1.5.3"
plotly = "^5.13.1"
nbformat = ">=4.2.0"

[tool.poetry.scripts]
scrape = "main:main"
backfill = "main:scrape_job_descriptions"


[tool.black]
line-length = 88
target-version = ["py311"]

[tool.isort]
profile = "black"

[tool.pylint.basic]
good-names = ["pk", "to", "_", "i", "n", "fp", "tz"]

[tool.pylint.typecheck]
signature-mutators = ["click.decorators.option"]
ignored-classes = ["sqlalchemy.orm.session.sessionmaker"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
